requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.4.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
pydantic-settings==2.1.0
asyncpg==0.29.0
flower==2.0.1